Refactored naming and aliased all imports
