+++ main.py
@@ -1,5 +1,5 @@
 from matrx.API import api
-from matrx.scenarios import simple_scenario, test_scenario, new_vis_test
+from matrx.use_cases import simple_case, test_case, new_vis_test
 
 if __name__ == "__main__":
