+++ matrx/actions/__init__.py
@@ -1,2 +1,4 @@
-
-
+from matrx.actions.action import Action
+from matrx.actions.door_actions import *
+from matrx.actions.move_actions import *
+from matrx.actions.object_actions import *
