+++ matrx/goals/__init__.py
@@ -1,0 +1,1 @@
+from matrx.goals.goals import *
