+++ matrx/grid_world.py
@@ -11,7 +11,7 @@ from matrx.actions.object_actions import *
 from matrx.logger.logger import GridWorldLogger
 from matrx.objects.env_object import EnvObject
 from matrx.objects.standard_objects import AreaTile
-from matrx.messages import MessageManager
+from matrx.messages.message_manager import MessageManager
 from matrx.objects.agent_body import _get_all_classes
 from matrx.API import api
