+++ matrx/logger/log_agent_actions.py
@@ -1,5 +1,4 @@
 from matrx.logger.logger import GridWorldLogger
-from matrx.grid_world import GridWorld
 
 
 class LogActions(GridWorldLogger):
@@ -8,7 +7,7 @@ class LogActions(GridWorldLogger):
         super().__init__(save_path=save_path, file_name=file_name_prefix, file_extension=file_extension,
                          delimiter=delimeter, log_strategy=1)
 
-    def log(self, grid_world: GridWorld, agent_data: dict):
+    def log(self, grid_world, agent_data):
         log_data = {}
         for agent_id, agent_body in grid_world.registered_agents.items():
             log_data[agent_id] = agent_body.current_action
