+++ matrx/logger/log_idle_agents.py
@@ -1,5 +1,4 @@
 from matrx.logger.logger import GridWorldLogger
-from matrx.grid_world import GridWorld
 
 
 class LogIdleAgents(GridWorldLogger):
@@ -8,7 +7,7 @@ class LogIdleAgents(GridWorldLogger):
         super().__init__(log_strategy=log_strategy, save_path=save_path, file_name=file_name_prefix,
                          file_extension=file_extension, delimiter=delimeter)
 
-    def log(self, grid_world: GridWorld, agent_data: dict):
+    def log(self, grid_world, agent_data):
         log_statement = {}
         for agent_id, agent_obj in grid_world.registered_agents.items():
             idle = agent_obj.properties['current_action'] is None
