+++ matrx/messages/__init__.py
@@ -1,0 +1,1 @@
+from matrx.messages.message import Message
