+++ matrx/objects/__init__.py
@@ -1,0 +1,3 @@
+from matrx.objects.agent_body import AgentBody
+from matrx.objects.env_object import EnvObject
+from matrx.objects.standard_objects import *
